Moved the method to SoftwareUpgrade.js file
